Remove from Cart

+++ ajaxCallbacks/delete_from_cart.php
@@ -5,9 +5,14 @@ ob_start();
 $json = array();
 $json['success'] = false;
 $id = $_POST['cart_key'];
-$query = "delete from cart where id=$id";
-$json['content'] = ob_get_clean();
-if(mysqli_query($conn, $query)){
+if(isset($_SESSION['id']) && $_SESSION['id'] != ''){
+    $query = "delete from cart where id=$id";
+    $json['content'] = ob_get_clean();
+    if(mysqli_query($conn, $query)){
+        $json['success'] = true;
+    }
+}else{
+    unset($_SESSION['cart'][$id]);
     $json['success'] = true;
 }
 echo json_encode($json);
